+++ deployer.go
@@ -2,7 +2,10 @@ package main
 
 import (
 	"fmt"
+	"log"
 	"os"
+	"os/exec"
+	"strings"
 
 	"github.com/progrium/go-basher"
 )
@@ -27,13 +30,60 @@ func BinVersion(args []string) {
 	fmt.Println(version())
 }
 
+// Copied from https://github.com/progrium/go-basher/blob/master/basher.go#L37
+// to be able to add exporting DEBUG and TRACE, as we don't inherit anything
+// from parent bash
+
+func application(
+	funcs map[string]func([]string),
+	scripts []string,
+	loader func(string) ([]byte, error),
+	copyEnv bool) {
+
+	var bashPath string
+	bashPath, err := exec.LookPath("bash")
+	if err != nil {
+		if strings.Contains(os.Getenv("SHELL"), "bash") {
+			bashPath = os.Getenv("SHELL")
+		} else {
+			bashPath = "/bin/bash"
+		}
+	}
+	bash, err := basher.NewContext(bashPath, os.Getenv("DEBUG") != "")
+
+	bash.Export("DEBUG", os.Getenv("DEBUG"))
+	bash.Export("TRACE", os.Getenv("TRACE"))
+
+	if err != nil {
+		log.Fatal(err)
+	}
+	for name, fn := range funcs {
+		bash.ExportFunc(name, fn)
+	}
+	if bash.HandleFuncs(os.Args) {
+		os.Exit(0)
+	}
+
+	for _, script := range scripts {
+		bash.Source(script, loader)
+	}
+	if copyEnv {
+		bash.CopyEnv()
+	}
+	status, err := bash.Run("main", os.Args[1:])
+	if err != nil {
+		log.Fatal(err)
+	}
+	os.Exit(status)
+}
+
 func main() {
 	if len(os.Args) == 2 && os.Args[1] == "--version" {
 		fmt.Println("CloudBreak Deployer:", version())
 		os.Exit(0)
 	}
 
-	basher.Application(map[string]func([]string){
+	application(map[string]func([]string){
 		"hello":       Hello,
 		"bin-version": BinVersion,
 	}, []string{